requirements_dev.txt
Django==1.6.5
Pillow==2.5.3
argparse==1.2.1
django-appconf==0.6
django-user-accounts==1.0c9
oauthlib==0.6.3
psycopg2==2.5.2
pytz==2014.4
six==1.8.0
wsgiref==0.1.2